requirements.txt
chris_plugin>=0.3.0
opencv-python
pydicom
# pflog==1.2.26
# pftel-client~=1.0.6
# for bug fix on transfer syntax errors
pylibjpeg
pylibjpeg-libjpeg
python-gdcm
# for running OCR
pillow
pytesseract